# docs/api/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspace": {
      "get": {
        "operationId": "WorkspaceController_all_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets all organization workspace resources",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/products": {
      "post": {
        "operationId": "WorkspaceController_createProduct_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates a product",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/products/{id}": {
      "patch": {
        "operationId": "WorkspaceController_updateProduct_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates a product",
        "tags": [
          "Workspace"
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_deleteProduct_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Deletes a product",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/events": {
      "post": {
        "operationId": "WorkspaceController_createEvent_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates an event",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/events/{id}": {
      "patch": {
        "operationId": "WorkspaceController_updateEvent_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an event",
        "tags": [
          "Workspace"
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_deleteEvent_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Deletes an event",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/custom-fields": {
      "post": {
        "operationId": "WorkspaceController_createField_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates a custom student field",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/custom-fields/{id}": {
      "patch": {
        "operationId": "WorkspaceController_updateField_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates a custom student field",
        "tags": [
          "Workspace"
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_deleteField_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Deletes a custom student field",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/student-field-values/{studentId}": {
      "patch": {
        "operationId": "WorkspaceController_replaceValues_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Replaces a student's custom field values",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/guardian-field-values/{guardianId}": {
      "patch": {
        "operationId": "WorkspaceController_replaceGuardianValues_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Replaces a guardian's custom field values",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/broadcasts": {
      "post": {
        "operationId": "WorkspaceController_createBroadcast_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates a broadcast message",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/broadcasts/{id}": {
      "patch": {
        "operationId": "WorkspaceController_updateBroadcast_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates a broadcast message",
        "tags": [
          "Workspace"
        ]
      },
      "delete": {
        "operationId": "WorkspaceController_deleteBroadcast_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Deletes a broadcast message",
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/v1/workspace/broadcast-sends": {
      "post": {
        "operationId": "WorkspaceController_queueBroadcast_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Queues a broadcast for selected students",
        "tags": [
          "Workspace"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mensaly API",
    "description": "Versioned backend API for Mensaly company owners and platform administrators",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mensaly_session"
      }
    },
...
  }
}